nights count wip
