+++ app/javascript/controllers/flatpickr_controller.js
@@ -1,0 +1,24 @@
+import { Controller } from 'stimulus';
+import { Application } from 'stimulus'
+import { definitionsFromContext } from 'stimulus/webpack-helpers'
+
+const application = Application.start()
+const context = require.context('../controllers', true, /\.js$/)
+application.load(definitionsFromContext(context))
+
+// import Flatpickr
+import Flatpickr from 'stimulus-flatpickr'
+
+// Import style for flatpickr
+require("flatpickr/dist/flatpickr.css")
+
+// Manually register Flatpickr as a stimulus controller
+application.register('flatpickr', Flatpickr)
+
+export default class extends Controller {
+  static targets = ['output'];
+
+  connect() {
+    console.log("123")
+  }
+}
