--- app/javascript/controllers/room_show_controller.js
@@ -1,7 +1,0 @@
-import { Controller } from 'stimulus';
-
-export default class extends Controller {
-  static targets = ['output'];
-
-  connect() {}
-}
