+++ app/models/room.rb
@@ -13,9 +13,9 @@ class Room < ApplicationRecord
 
   enum room_type: [:entire_place, :private_rooms, :hotel_rooms, :shared_rooms]
 
-  enum checkin_start_at: {"15:00": 15, "11:00": 11, "12:00": 12, "13:00": 13, "14:00": 14, "16:00": 16, "17:00": 17, "18:00": 18}, _prefix: :checkin_start_at
-  enum checkin_end_at: { "半夜": 24, "14:00": 14, "15:00": 15, "16:00": 16, "17:00": 17, "18:00": 18, "19:00": 19, "20:00": 20, "21:00": 21, "22:00": 22, "23:00": 23}, _prefix: :checkin_end_at
-  enum checkout_time: {"12:00": 12, "11:00": 11, "13:00": 13, "14:00": 14, "15:00": 15, "16:00": 16}, _prefix: :checkout_time
+  enum checkin_start_at: {"11:00": 11, "12:00": 12, "13:00": 13, "14:00": 14, "15:00": 15, "16:00": 16, "17:00": 17, "18:00": 18}, _prefix: :checkin_start_at
+  enum checkin_end_at: {"14:00": 14, "15:00": 15, "16:00": 16, "17:00": 17, "18:00": 18, "19:00": 19, "20:00": 20, "21:00": 21, "22:00": 22, "23:00": 23, "半夜": 24}, _prefix: :checkin_end_at
+  enum checkout_time: {"11:00": 11, "12:00": 12, "13:00": 13, "14:00": 14, "15:00": 15, "16:00": 16}, _prefix: :checkout_time
 
   scope :not_deleted, -> { where(deleted_at: nil) }
 end
