+++ app/views/rooms/_form.html.erb
@@ -94,11 +94,11 @@
       <div class="flex gap-5 font-midium">
         <div class="flex items-center gap-2 w-content">
           <p class="font-medium">從</p>
-          <%= room.select :checkin_start_at, options_for_select(Room.checkin_start_ats.keys), {}, class:"input"%>
+          <%= room.select :checkin_start_at, options_for_select(Room.checkin_start_ats.keys, '15:00'), {}, class:"input"%>
         </div>
         <div class="flex items-center gap-2 w-content ">
           <p class="font-medium">到</p>
-          <%= room.select :checkin_end_at, options_for_select(Room.checkin_end_ats.keys), {}, class:"input"%>
+          <%= room.select :checkin_end_at, options_for_select(Room.checkin_end_ats.keys, '半夜'), {}, class:"input"%>
         </div>
       </div>
     </div>
@@ -107,7 +107,7 @@
         <p class="text-sm text-red-700 align-middle ">*</p>
         <%= room.label :checkout_time, "退房時間" %>
       </div>
-      <%= room.select :checkout_time, options_for_select(Room.checkout_times.keys), {}, class:"input", selected:"14"%>
+      <%= room.select :checkout_time, options_for_select(Room.checkout_times.keys, '12:00'), {}, class:"input", selected:"14"%>
     </div>
     <div class="flex items-center gap-2">
       <%= room.label :photos, '上傳圖片', class:"font-medium"%>
