+++ app/views/rooms/index.html.erb
@@ -8,7 +8,7 @@
               <i class="self-end text-2xl text-center fa-solid fa-heart opacity-60 hover:opacity-80" ></i>
               <div class="flex self-center justify-between w-full my-auto ">
                 <div class="flex items-center justify-center w-8 h-8 bg-white rounded-full opacity-50 hover:opacity-100 "
-              data-action="carousel#prev">
+              data-action="click->carousel#prev">
                   <i class="text-gray-700 fa-solid fa-chevron-left"></i>
                 </div>
                 <div class="flex items-center justify-center w-8 h-8 bg-white rounded-full opacity-50 hover:opacity-100 "
