+++ db/migrate/20221215063124_add_head_column_to_bookings.rb
@@ -1,0 +1,5 @@
+class AddHeadColumnToBookings < ActiveRecord::Migration[6.1]
+  def change
+    add_column :bookings, :headcount, :integer
+  end
+end
